重构yolo worker
